requirements.txt
Flask==2.0.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
yfinance==0.2.33
flask-sqlalchemy==3.1.1
flask-login==0.6.3
python-jose==3.3.0
Werkzeug==3.0.1
gunicorn==20.1.0
Flask-WTF==1.2.1
email-validator==2.1.0.post1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
beautifulsoup4
tradingview-ta
Flask-APScheduler
backoff==2.2.1
redis>=5.0.0
requests-html
lxml_html_clean
googlesearch-python
forex-python
wikipedia
duckduckgo_search
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
Flask-Limiter
redis[rate-limit]
Flask-WTF
pymysql
mysqlclient
Flask-Uploads==0.2.1
Pillow>=10.2.0
python-magic==0.4.27
python-magic-bin
google-generativeai>=0.3.2
google-ai-generativelanguage
google-cloud-aiplatform
pillow-heif
Flask-Compress==1.14
pyotp
qrcode
psycopg2
psycopg2-binary
flask-turnstile
gevent==21.8.0
gevent-websocket==0.10.1
hmac
stripe>=5.0.0